# [FIX] purchase_stock: improve partner._compute_on_time_rate perfs