Improve `res.partner._compute_on_time_rate` performances by calling
`mapped('quantity_done')` once on all the `order_lines.move_ids`.

Moving it out of the for loop allows to befinit from the performance
gains of the read_group call in `stock.move._quantity_done_compute`.

##### speedup

Client DB with 10k purchase.order.line and 111k stock_moves.
`res_partner._compute_on_time_rate` time by purchase.order.line.

| purchase.order.line | before PR | after PR |
|:-------------------:|:-----------:|:--------:|
| 20 | 0.091s | 0.055s |
| 80 | 0.041s | 0.040s |
| 175 | 0.125s | 0.057s |
| 904 | 0.322s | 0.086s |
| 7982 | 7.794s | 0.777s |


--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr