@Aurelienvd 2021-06-16
@ryv-odoo I restored the `filtered(lamba l: l.product_id.sudo().product_tmpl_id.type)` line to avoid breaking denis commit. It only slightly slows down the computation (with 7982 going from 680ms to 770ms) probably thanks to the `state` being in the search domain as you mentioned.